+++ mm/oom/adaptive-oom-score.c
@@ -1,0 +1,6 @@
+#include "adaptive-oom-score.skel.h"
+
+int main(void)
+{
+	return 0;
+}