+++ mm/oom/adaptive-oom-score.h
@@ -1,0 +1,6 @@
+#pragma once
+
+struct pf_event_t {
+	int pid;
+	unsigned long error_code;
+};